+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Reduce redundant source-memory and attention-cache copies during SMILES Transformer beam search ([#38](https://github.com/microsoft/retrochimera/pull/38)) ([@gabrieldernbach])
+- Reuse workers for SMILES canonicalization without ([#39](https://github.com/microsoft/retrochimera/pull/39)).
+
 ## [1.3.0] - 2026-09-21
 
 ### Changed
@@ -68,3 +73,4 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 [@kmaziarz]: https://github.com/kmaziarz
 [@lgeiger]: https://github.com/lgeiger
+[@gabrieldernbach]: https://github.com/gabrieldernbach